выполняет задание 5.14 Модуляция
