+++ js/functions.js
@@ -1,12 +1,5 @@
-
-
 // функция для проверки строки
 const checkLength = (string, lengthMax) => string.length <= lengthMax;
-
-сonsole.log(checkLength('стронгjhtgyu', 32));
-
-
-
 
 // функция для проверки на палидром
 const isPalindrome = (string) => {
@@ -18,9 +11,5 @@ const isPalindrome = (string) => {
   for(let i = string.length - 1; i >= 0; i--) {
     reversed += string[i];
   }
-
   return string === reversed;
 };
-
-console.log(isPalindrome('топот'));
-
